Add files via upload
Gym membership management system is a simple console application using linked list in C programming language. User can perform basic gym membership management operations like adding the membership, canceling the membership, searching for a particular member of the gym based on certain filters and displaying the members of the gym with their details.
Each member of the gym has a unique identification through their contact info and email id. The visitor can avail the membership of the gym by paying for the membership according to the period of membership they avail. It is convenient for the gym manager to see whose membership ends a particular month, so they can cancel or renew the membership. 
Data Structures used:
1)	Structures
2)	Files 
3)	Linked List